http: create separate source files for THttpWSEngine and THttpWSHandler
