+++ gui/webdisplay/v7/src/TWebWindow.cxx
@@ -19,6 +19,8 @@
 
 #include "THttpEngine.h"
 #include "THttpCallArg.h"
+#include "THttpWSEngine.h"
+#include "THttpWSHandler.h"
 
 #include <cassert>
 #include <cstring>
